# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: Flow Content Factory REST API
  description: >-
    Centralized persistent production memory and management API for Speak English With Flow.
    Enables Custom GPT agents, automated production pipelines, and human creators to interact with
    the editorial planning, script versions, duration, metadata, assets, and audit timeline.
  version: 2.0.0
  contact:
    name: Speak English With Flow Engineering
servers:
  - url: http://localhost:3000/api/v1
    description: Local Development Server
  - url: https://flow-content-factory.vercel.app/api/v1
    description: Production Server
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /episodes:
    get:
      summary: List and Filter Planning Episodes
      operationId: listEpisodes
      parameters:
        - name: playlist
          in: query
          schema:
            type: string
        - name: status
          in: query
          schema:
            type: string
        - name: search
          in: query
          schema:
            type: string
      responses:
        '200':
          description: List of episodes with current status.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
  /episodes/search:
    get:
      summary: Multi-criteria Episode Search
      operationId: searchEpisodes
      parameters:
        - name: q
          in: query
          schema:
            type: string
        - name: code
          in: query
          schema:
            type: string
        - name: playlist
          in: query
          schema:
            type: string
      responses:
        '200':
          description: Matching episodes.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiResponse'
...
  /planning/extend:
    post:
      summary: Extend Planning with New Episodes
      operationId: extendPlanning
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlanningExtendPayload'
      responses:
        '201':
          description: Planning extended.
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
    ApiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
  schemas:
    ApiResponse:
      type: object
      properties:
        success:
          type: boolean
        data:
          type: object
        meta:
          type: object
        error:
          type: object
          properties:
            code:
              type: string
            message:
              type: string
...
    PlanningExtendPayload:
      type: object
      required:
        - episodes
      properties:
        episodes:
          type: array
          items:
            type: object
            required:
              - codeSerie
              - title
              - conceptPlaylist
            properties:
              codeSerie:
                type: string
              title:
                type: string
              conceptPlaylist:
                type: string
              thumbnailText:
                type: string
              thumbnailVisual:
                type: string
              hook:
                type: string
              keywords:
                type: string
              description:
                type: string